pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastmind"
version = "0.1.2"
description = "A lightweight, event-driven multi-agent framework for embodied AI systems"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0"
authors = [
    { name = "xiefujin", email = "xiefujin@example.com" }
]
keywords = ["ai", "agent", "llm", "embodied-ai", "graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

examples = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

all = [
    "fastmind[dev,examples]",
]

[project.urls]
Homepage = "https://github.com/kandada/fastmind"
Documentation = "https://github.com/kandada/fastmind#readme"
Repository = "https://github.com/kandada/fastmind"

[tool.hatch.build.targets.wheel]
packages = ["."]
package-dir = "."

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
